requirements.txt
MLB-StatsAPI>=1.7.2
requests>=2.31.0
python-dotenv>=1.0.1
atproto>=0.0.55
